19-grok-self-coach/requirements.txt
pandas
chromadb
sentence-transformers
openai
python-dotenv
